+++ backend/src/main/java/com/alper/shotify/backend/repository/IRecommendationRepository.java
@@ -3,5 +3,5 @@ package com.alper.shotify.backend.repository;
 import com.alper.shotify.backend.entity.RecommendationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRecommendationRepository extends JpaRepository<RecommendationEntity, Long> {
+public interface IRecommendationRepository extends JpaRepository<RecommendationEntity, Integer> {
 }
