+++ backend/src/main/java/com/alper/shotify/backend/repository/ISongRepository.java
@@ -3,5 +3,5 @@ package com.alper.shotify.backend.repository;
 import com.alper.shotify.backend.entity.SongEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ISongRepository extends JpaRepository<SongEntity, Long> {
+public interface ISongRepository extends JpaRepository<SongEntity, Integer> {
 }
